Added navigator.vendor evasion
