+++ packages/puppeteer-extra-plugin-stealth/evasions/navigator.vendor/package.json
